test_ns2/requirements.txt
pytest
paho-mqtt
pysmbclient
